requirements.txt
# Base
numpy
torch
torchvision
wandb
tqdm
pillow
matplotlib
opencv-python
PyYAML
requests
scipy
thop

# Web app
Flask
Flask-Cors
pyngrok
Pyrebase4
python-dotenv
